+++ src/Core/Dfe.Complete.Application/KeyContacts/Queries/GetKeyContactsForProject.cs
@@ -5,28 +5,19 @@ using Dfe.Complete.Application.Projects.Interfaces;
 using Dfe.Complete.Domain.ValueObjects;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
 
 namespace Dfe.Complete.Application.KeyContacts.Queries;
 
-public record GetKeyContactsForProjectQuery(ProjectId ProjectId) : IRequest<Result<KeyContactDto>>;
+public record GetKeyContactsForProjectQuery(ProjectId ProjectId) : IRequest<Result<KeyContactDto?>>;
 
-public class GetKeyContactsForProjectQueryHandler(IKeyContactReadRepository keyContactsRepository, ILogger<GetKeyContactsForProjectQueryHandler> logger, IMapper mapper)
-    : IRequestHandler<GetKeyContactsForProjectQuery, Result<KeyContactDto>>
+public class GetKeyContactsForProjectQueryHandler(IKeyContactReadRepository keyContactsRepository, IMapper mapper)
+    : IRequestHandler<GetKeyContactsForProjectQuery, Result<KeyContactDto?>>
 {
-    public async Task<Result<KeyContactDto>> Handle(GetKeyContactsForProjectQuery request, CancellationToken cancellationToken)
+    public async Task<Result<KeyContactDto?>> Handle(GetKeyContactsForProjectQuery request, CancellationToken cancellationToken)
     {
-        try
-        {
-            var keyContacts = await keyContactsRepository.KeyContacts.FirstOrDefaultAsync(contact => contact.ProjectId != null && contact.ProjectId == request.ProjectId,
-                    cancellationToken);
+        var keyContacts = await keyContactsRepository.KeyContacts.FirstOrDefaultAsync(contact => contact.ProjectId != null && contact.ProjectId == request.ProjectId,
+                cancellationToken);
 
-            return Result<KeyContactDto>.Success(keyContacts == null ? new KeyContactDto() : mapper.Map<KeyContactDto>(keyContacts));
-        }
-        catch (Exception ex)
-        {
-            logger.LogError(ex, "Exception for {Name} Request - {@Request}", nameof(GetKeyContactsForProjectQueryHandler), request);
-            return Result<KeyContactDto>.Failure(ex.Message);
-        }
+        return Result<KeyContactDto?>.Success(mapper.Map<KeyContactDto>(keyContacts));
     }
 }